github-agent/requirements.txt
google-generativeai
requests
python-dotenv
flask
markdown
